scanner/requirements.txt
Pillow~=8.3.2
Flask~=2.0.2
pyzbar~=0.1.8
